requirements.txt
streamlit
llama-cpp-python
faster-whisper
pypdf
python-docx
pyttsx3
Pillow
huggingface_hub
streamlit-audiorecorder
numpy
soundfile
